feat: implement capacity analyzer for saturation guardrails
Add fast, reactive capacity analyzer that prevents capacity exhaustion
by monitoring live vLLM metrics (KV cache usage and request queue length).

**Key Features:**
- Operates from live vLLM Prometheus metrics (no offline profiling)
- Detects imminent capacity exhaustion across all variants of a model
- Implements spare capacity calculations and worst-case scale-down safety simulation
- Integrates with demand-driven optimization via hybrid decision matrix
- Provides capacity safety overrides to prevent system overload

**Implementation:**
- CapacityAnalyzer interface with analysis and arbitration methods
- Analyzer implements 5-step algorithm:
  1. Identify non-saturated replicas (below KV cache and queue thresholds)
  2. Calculate spare capacity per replica
  3. Average spare capacity across non-saturated replicas
  4. Determine scale-up recommendation (spare < trigger thresholds)
  5. Simulate scale-down safety with worst-case load redistribution

- Metrics collector (internal/collector/capacity_metrics.go) queries Prometheus:
  - Uses max_over_time[1m] for safety-first capacity analysis
  - Captures peak values: vllm:kv_cache_usage_perc and vllm:num_requests_waiting
  - Prevents missing saturation events that could occur between instant queries

- Multi-variant aggregation: analyzes capacity across all variants
  of the same model, providing both aggregated and per-variant breakdown

**Hybrid Decision Matrix:**
- Capacity veto: prevents scale-down when capacity needs more
- Safety block: blocks unsafe scale-down
- Demand-driven: allows demand decisions when capacity is safe
- Capacity-driven: triggers scale-up when spare capacity is low

**Testing:**
- Comprehensive unit tests (8 test cases, all passing)
- Test coverage: scale-up triggers, scale-down safety, multi-variant
  aggregation, arbitration matrix, saturated replica identification

**Configuration:**
Uses capacity scaling thresholds from ConfigMap:
- kvCacheThreshold: 0.80 (saturation at ≥80%)
- queueLengthThreshold: 5 (saturation at ≥5 queued)
- kvSpareTrigger: 0.1 (scale-up when avg spare <10%)
- queueSpareTrigger: 3 (scale-up when avg spare <3)

Related: GitHub issue #269

+++ internal/interfaces/capacity_scaling.go
@@ -16,13 +16,13 @@ type CapacityScalingConfig struct {
 	KvCacheThreshold float64 `yaml:"kvCacheThreshold"`
 
 	// QueueLengthThreshold: Replica is saturated if queue length >= this threshold
-	QueueLengthThreshold int `yaml:"queueLengthThreshold"`
+	QueueLengthThreshold float64 `yaml:"queueLengthThreshold"`
 
 	// KvSpareTrigger: Scale-up if average spare KV cache capacity < this value (0.0-1.0)
 	KvSpareTrigger float64 `yaml:"kvSpareTrigger"`
 
 	// QueueSpareTrigger: Scale-up if average spare queue capacity < this value
-	QueueSpareTrigger int `yaml:"queueSpareTrigger"`
+	QueueSpareTrigger float64 `yaml:"queueSpareTrigger"`
 }
 
 // DefaultCapacityScalingConfig returns hardcoded default configuration.
@@ -61,13 +61,13 @@ func (c *CapacityScalingConfig) Validate() error {
 		return fmt.Errorf("kvCacheThreshold must be between 0 and 1, got %.2f", c.KvCacheThreshold)
 	}
 	if c.QueueLengthThreshold < 0 {
-		return fmt.Errorf("queueLengthThreshold must be >= 0, got %d", c.QueueLengthThreshold)
+		return fmt.Errorf("queueLengthThreshold must be >= 0, got %.1f", c.QueueLengthThreshold)
 	}
 	if c.KvSpareTrigger < 0 || c.KvSpareTrigger > 1 {
 		return fmt.Errorf("kvSpareTrigger must be between 0 and 1, got %.2f", c.KvSpareTrigger)
 	}
 	if c.QueueSpareTrigger < 0 {
-		return fmt.Errorf("queueSpareTrigger must be >= 0, got %d", c.QueueSpareTrigger)
+		return fmt.Errorf("queueSpareTrigger must be >= 0, got %.1f", c.QueueSpareTrigger)
 	}
 	// KV cache threshold should be greater than spare trigger (otherwise contradictory)
 	if c.KvCacheThreshold < c.KvSpareTrigger {
